Add files via upload
Pickling in python

Pickling is the process of converting a Python object hierarchy into a byte stream. It is used for serializing and de-serializing a Python object structure. This process is also called “serialization”, “marshalling”, or “flattening”. The byte stream can then be transmitted or stored, and later reconstructed to create a new object with the same structure.